Version 3.0.0

+++ service/src/main/java/org/grnet/cat/services/registry/TestService.java
@@ -1,43 +1,46 @@
 package org.grnet.cat.services.registry;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.quarkus.hibernate.orm.panache.Panache;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.ForbiddenException;
+import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.UriInfo;
 import org.grnet.cat.dtos.pagination.PageResource;
-import org.grnet.cat.dtos.registry.principle.PrincipleResponseDto;
 import org.grnet.cat.dtos.registry.test.*;
-import org.grnet.cat.entities.registry.Principle;
 import org.grnet.cat.entities.registry.Test;
-import org.grnet.cat.entities.registry.TestDefinition;
+import org.grnet.cat.entities.registry.TestMethod;
 import org.grnet.cat.exceptions.UniqueConstraintViolationException;
 import org.grnet.cat.mappers.registry.MotivationMapper;
-import org.grnet.cat.mappers.registry.PrincipleMapper;
 import org.grnet.cat.mappers.registry.TestMapper;
+import org.grnet.cat.repositories.MotivationAssessmentRepository;
 import org.grnet.cat.repositories.registry.MetricTestRepository;
-import org.grnet.cat.repositories.registry.TestDefinitionRepository;
+import org.grnet.cat.repositories.registry.TestMethodRepository;
 import org.grnet.cat.repositories.registry.TestRepository;
 
 import org.jboss.logging.Logger;
 
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @ApplicationScoped
 public class TestService {
 
     @Inject
+    ObjectMapper objectMapper;
+    @Inject
     TestRepository testRepository;
     @Inject
     MetricTestRepository metricTestRepository;
 
     @Inject
-    TestDefinitionService testDefinitionService;
+    TestMethodRepository testMethodRepository;
 
     @Inject
-    TestDefinitionRepository testDefinitionRepository;
+    MotivationAssessmentRepository motivationAssessmentRepository;
 
     private static final Logger LOG = Logger.getLogger(TestService.class);
 
@@ -57,66 +60,46 @@ public class TestService {
 
 
     /**
-     * Retrieves a specific Test item and related Test Definition item by its ID.
-     *
-     * @param id The unique ID of the Test item.
-     * @return The corresponding Test DTO.
-     */
-    public TestAndTestDefinitionResponse getTestAndTestDefinitionById(String id) {
-
-        var test = testRepository.findById(id);
-        var testDefinition = testDefinitionRepository.fetchTestDefinitionByTestId(id);
-
-        return testResponseWithMotivations(test, testDefinition);
-    }
-
-
-
-
-    /**
      * This method takes a Principle entity, converts it to a PrincipleResponseDto, retrieves and maps
      * any associated motivations, and then sets the motivations in the response.
      *
      * @param test The Test entity to be converted and enhanced.
      * @return A PrincipleResponseDto with associated motivations.
      */
-    private TestAndTestDefinitionResponse testResponseWithMotivations(Test test, TestDefinition testDefinition) {
+    private TestResponseDto testResponseWithMotivations(Test test) {
 
-        var testAndTestDefinitionResponse = TestMapper.INSTANCE.testAndTestDefinitionToDto(test, testDefinition);
+        var testResponse = TestMapper.INSTANCE.testToDto(test);
 
         var motivations = testRepository.getMotivationIdsByTest(test.getId());
         var motivationResponses = motivations.stream()
                 .map(MotivationMapper.INSTANCE::mapPartialMotivation)
                 .collect(Collectors.toList());
 
-        testAndTestDefinitionResponse.setMotivations(motivationResponses);
+        testResponse.setMotivations(motivationResponses);
 
-        return testAndTestDefinitionResponse;
+        return testResponse;
     }
 
 
-
-    /**
-     * Creates a new Test item.
-     *
-     * @param userId         The user creating the Test.
-     * @param request The Test and Test Definition request data.
-     * @return The created Test DTO.
-     */
     @Transactional
-    public TestAndTestDefinitionResponse createTestAndTestDefinition(String userId, TestAndTestDefinitionRequest request) {
+    public TestResponseDto createTest(String userId, TestRequestDto request) {
 
-        if (testRepository.notUnique("TES", request.getTestRequest().TES.toUpperCase())) {
-            throw new UniqueConstraintViolationException("tes", request.getTestRequest().TES.toUpperCase());
+        if (testRepository.notUnique("TES", request.TES.toUpperCase())) {
+            throw new UniqueConstraintViolationException("tes", request.TES.toUpperCase());
         }
 
-        var test = TestMapper.INSTANCE.testToEntity(request.getTestRequest());
+        var test = TestMapper.INSTANCE.testToEntity(request);
+
+        test.setTestMethod(Panache.getEntityManager().getReference(TestMethod.class, request.testMethodId));
         test.setPopulatedBy(userId);
+        //test.setTestParams(request.testParams);
+        test.setVersion(1);
+
         testRepository.persist(test);
+        test.setLodTES_V(test.getId());
 
-        var testDefinition = testDefinitionService.createTestDefinition(userId, request.getTestDefinitionRequest(), test.getId());
 
-        return TestMapper.INSTANCE.testAndTestDefinitionToDto(test, testDefinition);
+        return TestMapper.INSTANCE.testToDto(test);
     }
 
     /**
@@ -125,10 +108,9 @@ public class TestService {
      * @param id      The unique ID of the Test to update.
      * @param userId  The user performing the update.
      * @param request The Test update data.
-     * @return The updated Test DTO.
      */
     @Transactional
-    public void updateTest(String id, String userId, TestAndTestDefinitionUpdateRequest request) {
+    public void updateTest(String id, String userId, TestUpdateDto request) {
 
         if (metricTestRepository.existTestInStatus(id, Boolean.TRUE)) {
             throw new ForbiddenException("No action permitted, test exists in a published motivation");
@@ -137,15 +119,22 @@ public class TestService {
         var test = testRepository.findById(id);
         test.setPopulatedBy(userId);
 
-        if(!Objects.isNull(request.getTestRequest())){
+        if (!Objects.isNull(request)) {
 
-            TestMapper.INSTANCE.updateTestFromDto(request.getTestRequest(), test);
-        }
+            if (!Objects.equals(test.getTES(), request.TES)) {
+                throw new ForbiddenException("The TES name must be the same.");
+            }
 
-        if(!Objects.isNull(request.getTestDefinitionRequest())){
+            TestMapper.INSTANCE.updateTestFromDto(request, test);
 
-            var testDefinition = testDefinitionRepository.fetchTestDefinitionByTestId(id);
-            testDefinitionService.updateTestDefinition(testDefinition.getId(), userId, request.getTestDefinitionRequest());
+            if(!Objects.isNull(request.testMethodId)){
+
+                testMethodRepository.findByIdOptional(request.testMethodId).orElseThrow(()-> new NotFoundException("There is no Test Method with the following id: "+request.testMethodId));
+                test.setTestMethod(Panache.getEntityManager().getReference(TestMethod.class, request.testMethodId));
+                test.setTestQuestion(request.testQuestion);
+                test.setTestParams(request.testParams);
+                test.setToolTip(request.tooltip);
+            }
         }
 
     }
@@ -157,30 +146,77 @@ public class TestService {
      */
     @Transactional
     public boolean deleteTest(String id) {
+
         if (metricTestRepository.existTestInStatus(id, Boolean.TRUE)) {
             throw new ForbiddenException("No action permitted, test exists in a published motivation");
+        }
+
+        var test = testRepository.findById(id);
+        if (test != null && test.getVersion() == 1 && testRepository.countVersion(test.getLodTES_V()) > 1) {
+            throw new ForbiddenException("Cannot delete version 1 of the test as more versions exist");
         }
 
         return testRepository.deleteById(id);
     }
 
-//    /**
-//     * Retrieves a page of Test items.
-//     *
-//     * @param page    The index of the page to retrieve (starting from 0).
-//     * @param size    The maximum number of Test items to include in a page.
-//     * @param uriInfo The Uri Info for pagination links.
-//     * @return A PageResource containing the Test items in the requested page.
-//     */
-//    public PageResource<TestResponseDto> getTestlistAll(int page, int size, UriInfo uriInfo) {
-//
-//        var testPage = testRepository.fetchTestByPage(page, size);
-//        var testDtos = TestMapper.INSTANCE.testToDtos(testPage.list());
-//
-//        return new PageResource<>(testPage, testDtos, uriInfo);
-//    }
+    @Transactional
+    public TestResponseDto versionTest(String id, String userId, TestVersionRequestDto request) {
 
-    public PageResource<TestAndTestDefinitionResponse> getTestAndTestDefinitionListAll(String search, String sort, String order, int page, int size, UriInfo uriInfo) {
+        var parentId = testRepository.findById(id).getLodTES_V();
+        var parentTestVersion = testRepository.countVersion(parentId);
+
+        var childTest = TestMapper.INSTANCE.versionTestToEntity(request);
+        childTest.setTES(testRepository.findById(id).getTES());
+        childTest.setPopulatedBy(userId);
+        childTest.setLodTES_V(testRepository.findById(id).getLodTES_V());
+        childTest.setTestMethod(Panache.getEntityManager().getReference(TestMethod.class, request.testMethodId));
+
+        var newVersion = (int) parentTestVersion + 1 ;
+
+        childTest.setVersion(newVersion);
+        testRepository.persist(childTest);
+
+        return TestMapper.INSTANCE.testToDto(childTest);
+    }
+
+    @Transactional
+    public boolean deleteTestAllVersions(String id) {
+
+        if (metricTestRepository.existTestInStatus(id, Boolean.TRUE)) {
+            throw new ForbiddenException("No action permitted, test exists in a published motivation");
+        }
+        var test = testRepository.findById(id).getLodTES_V();
+
+        return testRepository.deleteAllVersions(test);
+    }
+
+    public List<TestResponseDto> getTestListAllVersions(String id, Integer latestVersion) {
+
+        var testParent = testRepository.findById(id);
+        if (testParent == null) {
+            throw new EntityNotFoundException("Test with ID " + id + " not found.");
+        }
+        var lodTES_V = testParent.getLodTES_V();
+
+        var testVersions = testRepository.fetchTestAllVersions(lodTES_V);
+
+        var testVersionsWithoutLatest = testVersions.stream()
+                .filter(test -> !test.getVersion().equals(latestVersion)) // Filter out the latest version
+                .collect(Collectors.toList());
+
+        if (testVersionsWithoutLatest.isEmpty()) {
+            return List.of();
+        }
+        var dtoList = testVersionsWithoutLatest.stream()
+                .map(this::testResponseWithMotivations)
+                .collect(Collectors.toList());
+
+        return dtoList;
+    }
+
+
+
+    public PageResource<TestResponseDto> getTestListAll(String search, String sort, String order, int page, int size, UriInfo uriInfo) {
 
         var testPage = testRepository.fetchTestByPage(search, sort, order, page, size);
         var tests = testPage.list();
@@ -188,23 +224,20 @@ public class TestService {
         if (tests.isEmpty()) {
             return new PageResource<>(testPage, List.of(), uriInfo);
         }
-        var testIds = tests.stream()
-                .map(Test::getId)
-                .collect(Collectors.toList());
-
-        var testDefinitions = testDefinitionRepository.fetchTestDefinitionsByTestIds(testIds);
-        var testDefinitionMap = testDefinitions.stream()
-                .collect(Collectors.toMap(TestDefinition::getLodTES, td -> td));
 
         var dtoList = tests.stream()
                 .map(test -> {
-                    TestDefinition testDefinition = testDefinitionMap.get(test.getId());
-                    return testResponseWithMotivations(test, testDefinition);
+                    var latestVersion = test.getVersion();
+                    var versions = getTestListAllVersions(test.getId(), latestVersion);
+
+                    var testResponse = testResponseWithMotivations(test);
+
+                    testResponse.setVersions(versions);
+
+                    return testResponse;
                 })
                 .collect(Collectors.toList());
 
         return new PageResource<>(testPage, dtoList, uriInfo);
     }
-
-
 }
